Submit full paths instead of drawing tiny lines
This consumes less CPU in the end by submitting less duplicated points
for rendering (and most paint engine tries to build a path in the end,
so let's avoid those conversions). In the case of SVG export is also
massively reduces the file size (a totally non scientific test
exhibited at least a factor 5 reduction of the file).